# fix: revert acl_winograd_convolution to stateful