# Description
partially reverts 16d6dd4f: "cpu: aarch64: Enable stateless ACL depthwise convolution"

reverts commit 03db3e4a: "cpu: aarch64: Call stateless ACL API from winograd convolution"

reverts commit 513f882d: "cpu: aarch64: hot fix for aux tensor management of stateless gemm-conv and winograd conv without lock."

Fixes #2324 
Fixes #2303

# Checklist

## General

- [x] Do all unit and benchdnn tests (`make test` and `make test_benchdnn_*`) pass locally for each commit?
- [x] Have you formatted the code using clang-format?

## Performance improvements
On a Neoverse V1 machine:

Current:

```
$ ./build-upstream/tests/benchdnn/benchdnn --max-ms-per-prb=3e3 --mode=P --conv --reset --allow-enum-tags-only=0 --engine=cpu --dir=FWD_I --alg=WINO --dt=f32:f32:f32 --stag=acdb --wtag=any --dtag=acdb --attr-scratchpad=user mb1_ic512oc512_ih7oh7kh3sh1dh0ph1_iw6ow6kw3sw1dw0pw1
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,wino:acl,,--mode=P --conv --allow-enum-tags-only=false --dir=FWD_I --stag=acdb --dtag=acdb --alg=wino --attr-scratchpad=user mb1ic512ih7iw6oc512oh7ow6kh3kw3ph1pw1,0.159384,1.25562,140.688,1.13288,147.603,1.07981
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):140.688 avg(ms):147.603
total: 3.32s; fill: 0.01s (0%);
```
After the revert:
```
$ ./build/tests/benchdnn/benchdnn --max-ms-per-prb=3e3 --mode=P --conv --reset --allow-enum-tags-only=0 --engine=cpu --dir=FWD_I --alg=WINO --dt=f32:f32:f32 --stag=acdb --wtag=any --dtag=acdb --attr-scratchpad=user mb1_ic512oc512_ih7oh7kh3sh1dh0ph1_iw6ow6kw3sw1dw0pw1        
Output template: perf,%engine%,%impl%,%name%,%prb%,%Gops%,%+ctime%,%-time%,%-Gflops%,%0time%,%0Gflops%
perf,cpu,wino:acl,,--mode=P --conv --allow-enum-tags-only=false --dir=FWD_I --stag=acdb --dtag=acdb --alg=wino --attr-scratchpad=user mb1ic512ih7iw6oc512oh7ow6kh3kw3ph1pw1,0.159384,1.66626,2.6106,61.0526,2.75155,57.925
tests:1 passed:1 skipped:0 mistrusted:0 unimplemented:0 invalid_arguments:0 failed:0 listed:0
total perf: min(ms):2.6106 avg(ms):2.75155
total: 3.26s; fill: 0.01s (0%);```
```