@theComputeKid 2025-01-08
@alvoron code freeze for oneDNN is end of the week, so we wanted to get in a stable version before then, this is the best we can do for now. We will try to test it a bit more and then get it in for Friday. Sorry for the slowdown.