@dzarukin 2025-01-09
In most cases it's most beneficial to create a dedicated revert commit to the one that it reverts. It's easier to track changes this way and re-apply patches in the future. Mixing all in one doesn't help with such flow.
Feel free to skip it this time but please refer to this practice in the future. Thanks.