@Radu2k 2025-01-09
> In most cases it's most beneficial to create a dedicated revert commit to the one that it reverts. It's easier to track changes this way and re-apply patches in the future. Mixing all in one doesn't help with such flow. Feel free to skip it this time but please refer to this practice in the future. Thanks.

Thanks @mgouicem @dzarukin, noted.

I assume that having 3 separate ones, one for each https://github.com/oneapi-src/oneDNN/commit/03db3e4ac043029b2b538b3eeaf843f1dde5b745 | https://github.com/oneapi-src/oneDNN/commit/513f882dd2fd348f2dba5ccdc344b515d7288df8 and a partially one for https://github.com/oneapi-src/oneDNN/commit/16d6dd4fdcdcf4c6a51b77764238c2eb97dad5f8, would have been ideal.